Recompute the "total sum" column upon rescaling
We rescaled the "total sum" column just like every other column.
Thereby, the "total sum" column need no longer be the total sum of the
other columns in that row (due to rounding errors, and missing values
in the rows to rescale).
Hence, we now recompute the "total sum" column upon rescaling and can
thereby guarantee that the "total sum" column of a rescaled row (i.e.:
for weekly, monthly, yearly) really is the total sum of the other
columns.

Change-Id: I5adf5fa5607730b08b7d45bdd21410b2f6c7a09e